Create 11.6.4.50pm
Passing Multidimensional Arrays
The most straightforward way of passing a multidimensional array to a function is to declare it in exactly the same way in the function as it was declared in the caller. If we were to call

	func(a2);
then we might declare

	func(int a[5][7])
	{
	...
	}
and it's clear that the array type which the caller passes is the same as the type which the function func accepts.

If we remember what we learned about simple arrays and functions, however, a question arises.  In our earlier function definitions, we were able to leave out the (single) array dimension, with the understanding that since the array was really defined in the caller, we didn't have to say (or know) how big it is. The situation is the same for multidimensional arrays, although it may not seem so at first.

The hypothetical function func above accepts a parameter a, where a is an array of 5 things, where each of the 5 things is itself an array. By the same argument that applies in the single-dimension case, the function does not have to know how big the array a is, overall. However, it certainly does need to know what a is an array of. It is not enough to know that a is an array of ``other arrays''; the function must know that a is an array of arrays of 7 ints. The upshot is that although it does not need to know how many ``rows'' the array has, it does need to know the number of columns. That is, if we want to leave out any dimensions, we can only leave out the first one:

	func(int a[][7])
	{
	...
	}
The second dimension is still required. (For a three- or more dimensional array, all but the first dimension are required; again, only the first dimension may be omitted.)

Here is a simple example of a function with a two dimensional array parameter that displays the element in position[1][1] (second row, second column) of that two dimensional array:

#include <iostream>
using namespace std;

void processArr(int a[][2]) {
   cout << "element at index 1,1 is " << a[1][1];
}
int main() {
   int arr[2][2];
   arr[0][0] = 0;
   arr[0][1] = 1;
   arr[1][0] = 2;
   arr[1][1] = 3;

   processArr(arr);
   return 0;
}

Output:

This will give the following output:

element at index 1,1 is 3
In the provided C++ program, the cout << endl; statement doesn't directly contribute to printing the 3x3 matrix. It serves a different purpose, which is to move the cursor to the next line after each row of the matrix is printed.

The actual printing of the matrix is handled by the inner for loop within the main() function. This loop iterates through each column of the TTT array, printing the corresponding character at each position using cout << TTT[r][c] << " ";. The space character (' ') after the array element ensures that each element is separated by a space.

Once the inner for loop completes, the cout << endl; statement is executed. This statement inserts a newline character, effectively moving the cursor to the next line. This is crucial for formatting the output as a matrix, as it ensures that each row of the matrix is printed on a separate line.

Without the cout << endl; statement, all the 'x' characters would be printed on a single line, resulting in a horizontal line of 'x's instead of the desired 3x3 matrix.

profile picture
WHAT IS THE CALLER IN THE STATEMENT....we were able to leave out the (single) array dimension, with the understanding that since the array was really defined in the caller,
View other drafts












The term "caller" in the statement "we were able to leave out the (single) array dimension, with the understanding that since the array was really defined in the caller" refers to the code that is passing the array to the function. In other words, the caller is the code that is responsible for creating and initializing the array, and then passing it to the function as an argument.

The reason why the caller can leave out the single array dimension when passing an array to a function is because the function only needs to know the base address of the array, not its size. The size of the array can be determined by the caller, and is passed to the function as a separate argument.

This is in contrast to multidimensional arrays, where the caller must specify all of the array dimensions when passing the array to a function. This is because multidimensional arrays are stored in memory in a more complex way, and the function needs to know all of the dimensions in order to access the elements of the array correctly.

Sources
www.eskimo.com/~scs/cclass/int/sx9a.html