Create 11.6.5.43pm
how is it possible to leave out the first array in a multidimensional array in c++ in...func(int a[][7])
View other drafts





























In C++, when you pass a multidimensional array to a function, you can omit the first dimension of the array. This is because arrays in C++ decay to pointers when passed to functions. A pointer holds the address of the first element in the array, so the function can access all of the elements of the array without knowing the size of the first dimension.

For example, consider the following code:

C++
void func(int a[][7]) {
  for (int i = 0; i < 5; i++) {
    for (int j = 0; j < 7; j++) {
      cout << a[i][j] << " ";
    }
    cout << endl;
  }
}

int main() {
  int a[5][7];
  for (int i = 0; i < 5; i++) {
    for (int j = 0; j < 7; j++) {
      a[i][j] = i * j;
    }
  }

  func(a);

  return 0;
}
Use code with caution. Learn more
In this code, the main() function creates a two-dimensional array a and initializes it with values. Then, it passes the array to the func() function. The func() function does not need to know the size of the first dimension of the array, because it can access all of the elements of the array using the pointer that is passed to it.

The reason why this works is because arrays in C++ are stored in memory in a contiguous block. This means that the elements of the array are stored in memory one after the other, with no gaps. As a result, the function can access the elements of the array by adding an offset to the base address of the array.

The offset for an element is calculated as follows:

C++
offset = (i * dimension2) + j
Use code with caution. Learn more
where:

i is the index of the element in the first dimension
dimension2 is the size of the second dimension
j is the index of the element in the second dimension
For example, to access the element at a[2][3], the function would calculate the offset as follows:

C++
offset = (2 * 7) + 3
Use code with caution. Learn more
This would give an offset of 17, which is the address of the element at a[2][3] in memory.

The ability to omit the first dimension of a multidimensional array when passing it to a function is a powerful feature of C++ that can make your code more concise and easier to read.